+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # BlogMore ChangeLog
 
+## v0.4.1
+
+**Unreleased**
+
+- Fixed relative URLs in RSS/Atom feed entry content being left as
+  root-relative paths; they are now rewritten to absolute URLs using the
+  configured site URL, resolving feed-validator warnings.
+  ([#107](https://github.com/davep/blogmore/pull/107))
+
 ## v0.4.0
 
 **Released: 2026-02-19**